+++ src/demo/components/ui/drawer/Drawer.tsx
@@ -1,0 +1,28 @@
+import React, { useCallback, useState } from "react";
+import { Drawer } from "../../../../components/ui/drawer/Drawer";
+
+interface Props {}
+
+const DrawerDemo: React.FC<Props> = (props) => {
+    // eslint-disable-next-line @typescript-eslint/no-shadow
+    const [open, setOpen] = useState(false);
+
+    const handleToggleOpen = useCallback(() => {
+        setOpen(p => !p);
+    }, []);
+
+    const handleClose = useCallback(() => {
+        setOpen(false);
+    }, []);
+
+    return (
+        <>
+            <Drawer isOpen={open} onClose={handleClose}>
+                Drawer content
+            </Drawer>
+            <button onClick={handleToggleOpen}>open drawer</button>
+        </>
+    );
+};
+
+export { DrawerDemo };